split main.go into multiple files based on their purpose
